% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm.R
\name{calc_glcm_edge}
\alias{calc_glcm_edge}
\title{Calculate edge for GLCM when processing block by block}
\usage{
calc_glcm_edge(shift, window)
}
\arguments{
\item{shift}{shift parameter for GLCM calculation}

\item{window}{window parameter for GLCM calculation}
}
\value{
vector of edge values
}
\description{
This is an internal function used to calculate edges for GLCM processing.
}
\keyword{internal}
