% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcm-package.R
\docType{data}
\name{test_raster}
\alias{test_raster}
\title{Randomly generated 100x100 test image}
\description{
Used in testing the output from the GLCM texture statistics C++ code.
}
\examples{
# The image was generated with the following code:
require(raster)
set.seed(0)
test_matrix <- matrix(runif(100)*32, nrow=10)
test_raster <- raster(test_matrix, crs='+init=epsg:4326')
test_raster <- cut(test_raster, seq(0, 32))
}
