% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{glcm-package}
\alias{glcm-package}
\title{Calculate textures from grey-level co-occurrence matrices (GLCMs) in R}
\description{
Enables calculation of image textures derived from grey-level co-occurrence
matrics (GLCMs) in R. The texture calculation is coded in C++ to optimize
computation time.
}
\author{
Alex Zvoleff, \email{azvoleff@conservation.org}
}
\keyword{package}

